rename `timer` module to `timing` to avoid name clash with Pylance's stuff
