+++ sandman_main/controls.py
@@ -6,7 +6,7 @@ Controls are used to manipulate parts of the bed.
 import enum
 import logging
 
-from sandman_main import timer
+import timing
 
 
 @enum.unique
